feat(*): upgrade to payload v2 (#109)
* chore(package.json): upgrade dependencies

* fix(jest.config.ts): add "transformIgnorePatterns"

* fix(jest.config.ts): try defining moduleDirectories

* fix(jest.config.ts): try removing transformIgnorePatterns

* test(jest.config.ts): try emulating config from payload

* test(package.json): try removing get-port

* test(jest.config.ts): try adjusting the ignore pattern

* test(jest.config.ts): try transforms again

* test(jest.config.ts): fix invalid config

* test(jest.config.ts): try a different preset

* refactor(*): stronger typing

* test(jest.config.ts): try a mongodb preset

* test(babel.config): try removing like in payload repo

* test(babel.config): try to fix es6 import error

* feat(*): upgrade to payload v2

* fix(tests): ensure tests have finished when closing connection

* fix(tests): try timeout beforeAll

* test(tests): increase timeout

* tests(tests): another timeout attempt

* test(tests): try another timeout approach

+++ src/api/payload-crowdin-sync/files.ts
@@ -16,9 +16,15 @@ import {
 } from "../helpers";
 import { isEmpty } from "lodash";
 
+import { CrowdinArticleDirectory } from '../../payload-types'
+
 export interface IcrowdinFile {
   id: string;
   originalId: number;
+  name: string;
+  field: string;
+  type: string;
+  updatedAt: string;
   fileData: {
     json?: Object;
     html?: string;
@@ -121,7 +127,7 @@ export class payloadCrowdinSyncFilesApi {
       const crowdinDirectory = await this.sourceFilesApi.createDirectory(
         this.projectId,
         {
-          directoryId: crowdinPayloadCollectionDirectory.originalId,
+          directoryId: crowdinPayloadCollectionDirectory.originalId as number,
           name: global ? collectionSlug : document.id,
           title: global
             ? toWords(collectionSlug)
@@ -415,9 +421,9 @@ export class payloadCrowdinSyncFilesApi {
     }
   }
 
-  async getArticleDirectory(documentId: string) {
+  async getArticleDirectory(documentId: string): Promise<CrowdinArticleDirectory | undefined> {
     const result = await getArticleDirectory(documentId, this.payload);
-    return result;
+    return result as CrowdinArticleDirectory | undefined;
   }
 
   async deleteFilesAndDirectory(documentId: string) {
@@ -434,6 +440,9 @@ export class payloadCrowdinSyncFilesApi {
     const crowdinPayloadArticleDirectory = await this.getArticleDirectory(
       documentId
     );
+    if (!crowdinPayloadArticleDirectory || !crowdinPayloadArticleDirectory.originalId) {
+      return
+    }
     await this.sourceFilesApi.deleteDirectory(
       this.projectId,
       crowdinPayloadArticleDirectory.originalId
